requirements.txt
# HHPDI 文档数据处理智能助手 — 依赖清单
# 安装命令：pip install -r requirements.txt

# ══════════════════════════════════════════════════════════════
#  桌面端（必需）
# ══════════════════════════════════════════════════════════════

# ── 工具①：文档解析（PDF/Word → MD）──────────────────────
pymupdf>=1.23.0        # PDF 解析
python-docx>=1.1.0     # Word 读取
Pillow>=10.0.0         # 图片处理
openai>=1.30.0         # VLM API 调用（兼容所有 OpenAI 格式供应商）

# ── 工具②：MD → Word 转换 ───────────────────────────────
# python-docx（已在上方）
beautifulsoup4>=4.12.0  # HTML 表格解析

# ── 工具③：MD 数据标注 ──────────────────────────────────
requests>=2.31.0        # LLM API 调用

# ── 可选：配置文件支持 ───────────────────────────────────
python-dotenv>=1.0.0

# ══════════════════════════════════════════════════════════════
#  REST API 服务（可选）
#  不安装时桌面程序正常运行，API 功能静默跳过。
#  安装后启动软件即自动开启 API，无需额外操作。
#  安装命令：pip install fastapi "uvicorn[standard]" python-multipart
# ══════════════════════════════════════════════════════════════

fastapi>=0.111.0           # REST API 框架，自动生成 OpenAPI / Swagger 文档
uvicorn[standard]>=0.29.0  # ASGI 服务器（websockets、httptools 加速版）
python-multipart>=0.0.9    # 文件上传支持（multipart/form-data）
